Review get_services.py into build_uuid_json.py
When you want to build the JSON listing all the Bluetooth defined UUIDs
you want to systematically use the raw source from bluetooth.org.

No need to use a class, which make the fact of using generators quite
useless.

If you need a definitions class, your cache is the UUID JSON this script
builds so the class should read and parse the UUID JSON and not some
HTML somewhere.